Adapt spawner-related hooks to async KubeSpawner
These changes mainly imply the transformation of hooks into
coroutines and the awaiting of calls to the kubernetes API to
read / write objects.

The post_stop_hooks can't be transformed into coroutines because
they are never awaited from the upstream spawner code, and they
need to be coroutines because they invoke the async k8s API.
Therefore, they have been removed from here and placed in an
override of the stop method in the SwanKubeSpawner, which is a
coroutine.